+++ stories/components/galactic/galactic.stories.ts
@@ -37,23 +37,63 @@ const Template: StoryFn<Partial<{ text: string, displayGalacticMenu: boolean }>>
     </wcs-galactic>
 `;
 const customContent = html`
-        <wcs-dropdown mode="clear" size="s">
-            <span slot="placeholder">Accessibilité</span>
-            <wcs-dropdown-header>Typographies</wcs-dropdown-header>
-            <wcs-dropdown-item @wcsDropdownItemClick="${e => e.stopPropagation()}">
-                <wcs-switch>Police dyslexie</wcs-switch>
-            </wcs-dropdown-item>
-            <wcs-dropdown-item @wcsDropdownItemClick="${e => e.stopPropagation()}">
-                <wcs-switch>Interlignage augmenté</wcs-switch>
-            </wcs-dropdown-item>
-            <wcs-dropdown-header>Constrastes</wcs-dropdown-header>
-           <wcs-dropdown-item @wcsDropdownItemClick="${e => e.stopPropagation()}">
-                <wcs-switch checked>Contrastes renforcés</wcs-switch>
-           </wcs-dropdown-item>
-        </wcs-dropdown>
+        <wcs-button id="accessibility-menu-button" mode="clear" size="s">
+            Accessibilité
+            <wcs-mat-icon size="s" icon="arrow_drop_down"></wcs-mat-icon>
+        </wcs-button>
+        <wcs-tooltip theme="light" trigger="click" interactive for="accessibility-menu-button" .appendTo=${() => document.querySelector("#tooltip-tippy")}>
+            <h3>Accessibilité</h3>
+            <wcs-switch>Police dyslexie</wcs-switch>
+            <wcs-switch>Interlignage augmenté</wcs-switch>
+            
+            <h3>Constrastes</h3>
+            <wcs-switch checked>Contrastes renforcés</wcs-switch>
+        </wcs-tooltip>
+        <div id="tooltip-tippy">
+        </div>
         <wcs-button mode="clear" size="s">Langue : FR</wcs-button>
 `;
 
+/**
+ * ## You want to add a button with (Show/Hide menu) Pattern
+ *
+ * Add an accessibility menu to configure parameters using `wcs-button` with `wcs-tooltip`.
+ *
+ * > **Important**: For correct tab navigation, you must add a `div` with id and set the `appendTo` property of the tooltip to target this div.
+ *
+ * Here's how to implement it:
+ *
+ * ```js
+ * // Lit example (adapt according to your framework - Angular, React)
+ * html`
+ *   <wcs-button id="accessibility-menu-button" mode="clear" size="s">
+ *     Accessibilité
+ *   </wcs-button>
+ *
+ *   <wcs-tooltip 
+ *     theme="wcs-primary" 
+ *     trigger="click" 
+ *     interactive 
+ *     for="accessibility-menu-button"
+ *     .appendTo=${() => document.querySelector("#tooltip-tippy")}
+ *   >
+ *     <h3>Accessibilité</h2>
+ *     <wcs-switch>Police dyslexie</wcs-switch>
+ *     <wcs-switch>Interlignage augmenté</wcs-switch>
+ *
+ *     <h3>Contrast</h2>
+ *     <wcs-switch checked>Contrastes renforcés</wcs-switch>
+ *   </wcs-tooltip>
+ *
+ *   <!-- Required for proper tab navigation ! -->
+ *   <div id="tooltip-tippy"></div>
+ *
+ *   <wcs-button mode="clear" size="s">
+ *     Language: FR
+ *   </wcs-button>
+ * `
+ * ```
+ */
 export const Default = Template.bind({});
 Default.args = {
     text: 'Un site'